Implement Integration test with docker container. Move classes from Core.kt
