+++ src/commonMain/kotlin/me/miguel/pgkn/PgKnException.kt
@@ -1,0 +1,3 @@
+package me.miguel.pgkn
+
+class SQLException : Exception()
