+++ src/commonMain/kotlin/me/miguel/pgkn/resultset/ResultSet.kt
@@ -1,153 +1,11 @@
-package me.miguel.pgkn
+package me.miguel.pgkn.resultset
 
-import kotlinx.cinterop.*
-import kotlinx.datetime.*
-import libpq.*
+import kotlinx.datetime.Instant
+import kotlinx.datetime.LocalDate
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.LocalTime
+import me.miguel.pgkn.SQLException
 
-fun PostgresDriver(
-    host: String,
-    port: Int = 5432,
-    database: String,
-    user: String,
-    password: String,
-): PostgresDriver = PostgresDriverImpl(
-    host = host,
-    port = port,
-    database = database,
-    user = user,
-    password = password
-)
-
-sealed interface PostgresDriver {
-    fun <T> executeQuery(sql: String, handler: (ResultSet) -> T): List<T>
-}
-
-private class PostgresDriverImpl(
-    host: String,
-    port: Int,
-    database: String,
-    user: String,
-    password: String,
-) : PostgresDriver {
-
-    private val connection = PQsetdbLogin(
-        pghost = host,
-        pgport = port.toString(),
-        dbName = database,
-        login = user,
-        pwd = password,
-        pgoptions = null,
-        pgtty = null
-    ).apply { require(ConnStatusType.CONNECTION_OK == PQstatus(this)) }!!
-
-    override fun <T> executeQuery(sql: String, handler: (ResultSet) -> T): List<T> = memScoped {
-        PQexecParams(
-            connection,
-            command = sql,
-            nParams = 0,
-            paramValues = createValues(0) {},
-            paramLengths = createValues(0) {},
-            paramFormats = createValues(0) {},
-            paramTypes = createValues(0) {},
-            resultFormat = TEXT_RESULT_FORMAT
-        )
-    }
-        .check()
-        .let {
-            val rs = PostgresResultSet(it)
-
-            val list: MutableList<T> = mutableListOf()
-            while (rs.next()) {
-                list.add(handler(rs))
-            }
-
-            PQclear(it)
-            return list
-        }
-
-    private fun CPointer<PGresult>?.check(): CPointer<PGresult> {
-        val status = PQresultStatus(this)
-        check(status == PGRES_TUPLES_OK || status == PGRES_COMMAND_OK || status == PGRES_COPY_IN) {
-            connection.error()
-        }
-        return this!!
-    }
-}
-
-
-private class PostgresResultSet(val internal: CPointer<PGresult>) : ResultSet {
-
-    private val rowCount: Int = PQntuples(internal)
-    private val columnCount: Int = PQnfields(internal)
-
-    private var currentRow = -1
-
-
-    override fun next(): Boolean {
-        if (currentRow > rowCount - 2) {
-            return false
-        }
-        currentRow++
-        return true
-    }
-
-    private fun isNull(columnIndex: Int): Boolean =
-        PQgetisnull(res = internal, tup_num = currentRow, field_num = columnIndex) == 1
-
-    private fun getPointer(columnIndex: Int): CPointer<ByteVar>? {
-        if (isNull(columnIndex)) return null
-        return PQgetvalue(res = internal, tup_num = currentRow, field_num = columnIndex) ?: throw SQLException()
-    }
-
-
-    /**
-     * Are all non-binary columns returned as text?
-     * https://www.postgresql.org/docs/9.5/libpq-exec.html#LIBPQ-EXEC-SELECT-INFO
-     */
-    override fun getString(columnIndex: Int): String? = getPointer(columnIndex)?.toKString().also { println(it) }
-
-    override fun getBoolean(columnIndex: Int): Boolean? = getString(columnIndex)?.equals("t")
-
-    override fun getShort(columnIndex: Int): Short? = getString(columnIndex)?.toShort()
-
-
-    override fun getInt(columnIndex: Int): Int? = getString(columnIndex)?.toInt()
-
-
-    override fun getLong(columnIndex: Int): Long? = getString(columnIndex)?.toLong()
-
-
-    override fun getFloat(columnIndex: Int): Float? = getString(columnIndex)?.toFloat()
-
-    override fun getDouble(columnIndex: Int): Double? = getString(columnIndex)?.toDouble()
-
-    override fun getBytes(columnIndex: Int): ByteArray? = getString(columnIndex)?.encodeToByteArray()
-
-    override fun getDate(columnIndex: Int): LocalDate? = getString(columnIndex)?.toLocalDate()
-
-    override fun getTime(columnIndex: Int): LocalTime? = getString(columnIndex)?.toLocalTime()
-    override fun getLocalDateTime(columnIndex: Int): LocalDateTime? = getString(columnIndex)
-        ?.fixIso8601()
-        ?.toLocalDateTime()
-
-    override fun getInstant(columnIndex: Int): Instant? = getString(columnIndex)
-        ?.fixIso8601()
-        ?.toInstant()
-}
-
-
-/**
- * To Fix ISO 8601, as postgres default is space not "T"
- * https://www.postgresql.org/docs/current/datatype-datetime.html#DATATYPE-DATETIME-OUTPUT
- */
-private fun String.fixIso8601() = replaceRange(10, 11, "T")
-
-private fun CPointer<PGconn>?.error(): String = PQerrorMessage(this)!!.toKString().also { PQfinish(this) }
-
-private const val TEXT_RESULT_FORMAT = 0
-private const val BINARY_RESULT_FORMAT = 1
-
-class SQLException : Exception()
 sealed interface ResultSet {
 
     /**
